Refactor to avoid circular dependencies
1. Split error handlers into layers so that lower level error handlers
   can be used separate from API error handlers
2. Split the read and write pieces of the personal access key functions
